pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dashtrash"
version = "0.1.0"
description = "Terminal-based dashboard for real-time monitoring. Real-time dashboards. Questionable aesthetics."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "dashtrash contributors", email = "info@dashtrash.dev"}
]
maintainers = [
    {name = "dashtrash contributors", email = "info@dashtrash.dev"}
]
keywords = ["dashboard", "monitoring", "terminal", "cli", "tui", "system", "logs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Terminals",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "pyfiglet>=0.8.0",
    "PyYAML>=6.0",
    "textual>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.991",
]

[project.urls]
Homepage = "https://github.com/turancannb02/dashtrash"
Documentation = "https://github.com/turancannb02/dashtrash#readme"
Repository = "https://github.com/turancannb02/dashtrash.git"
"Bug Tracker" = "https://github.com/turancannb02/dashtrash/issues"

[project.scripts]
dashtrash = "dashtrash.main:main"

[tool.setuptools]
packages = ["dashtrash", "dashtrash.panels", "dashtrash.plugins"]

[tool.setuptools.package-data]
dashtrash = ["*.yml", "*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true 